create timer-prompt component folder
